pyproject.toml
[tool.poetry]
name = "silent-synapses"
version = "0.1.0"
description = "Silent synapses python code"
authors = ["Andrew Li <andrewli@u.northwestern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "1.22.4"
boto3 = "^1.28.85"
scikit-learn = "^1.3.2"
matplotlib = "^3.8.1"
networkx = "^3.2.1"
plotly = "^5.18.0"
trimesh = "^4.0.4"
scipy = "^1.11.3"
tqdm = "^4.66.1"
skeletor = "^1.2.3"
nglui = "^2.14.3"
caveclient = "^5.13.0"
pandas = "1.5.1"
python-dotenv = "1.0.0"
# cloud-volume = "8.28.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
